MEN-1011 Add callback for reading state scripts.
While artifact is read, a callback will be called for the client
to process the scripts. The callback is returning an io.Reader
for the client to read the content of the script from, and os.FileInfo
interface for reading additional data (like name and size of the script).

Please note that there is one call made for each script.

Changelog: None

Signed-off-by: Marcin Pasinski <marcin.pasinski@mender.io>