Fix overwriting artifact if new one is invalid.
If some error occurs while creating new artifact the one existing
with the same name as newly created will be overwritten. This fix
is changing the library functionality so that file suffixed
with `.tmp` will be created first and the suffix will be removed only
if the new artifact is valid.

Changelog: Title

Signed-off-by: Marcin Pasinski <marcin.pasinski@mender.io>