MEN-1412 Fix broken header checksum verification.
For reading and verifying header checksum we are using checksum.Reader.
Internal inplementation of checksum.Reader is using io.TeeReader for
calculating checksum and reading the stream at the same time.
After receiving EOF checksum should be checked automatically inside
checksum.Reader, but if we are inside tar archive error is not passed up
the stack to tar.Reader.
The workaround is to check the sum again after the whole header is read.

Changelog: Title


Signed-off-by: Marcin Pasinski <marcin.pasinski@mender.io>