Refactor signing existing artifact and update tests.
For modifying artifact functionality we will need some parts
of the code to be shared with exiting signing artifact functionality.
This commit is the foundation for reusing the code with
modifying exiting artifact.

What is more, some tests for signing existing artifacts are added to
cover signing artifact containing scripts.

Changelog: None

Signed-off-by: Marcin Pasinski <marcin.pasinski@northern.tech>