Move all CLI related files to separate `cli/mender-artifact` folder.
This tool needs some refactoring as the initial idea of making this
only library for reading and writing an artifact is no longer the case.
Using it as a standalone tool for manipulating artifacts is a
significant part of the functionality and needs to be refactored
before we can add more changes.

This is the initial commit moving all the standalone part to the separate
folder and changing Makefile to build the tool as it was before.

Changelog: None

Signed-off-by: Marcin Pasinski <marcin.pasinski@northern.tech>