MEN-2155 Fail validate when key supplied for unsigned artifact
Changelog: A command of the form
"mender-artifact validate unsigned.mender -k public.key"
was incorrectly succeeding for an unsigned artifact when a public key
was supplied. Supplying a public key indicates that the caller requires
the artifact to contain a signature that matches that key.
Now this command fails (exits with a nonzero value) as expected.

Fixed logic in validate function to return an error when
a public key was given but we never received a callback to
validate a signature inside the artifact.

Added 2 more unit tests:

- Must fail validation of an unsigned artifact with key given.
  Confirmed this test failed before the fix mentioned above,
  and now passes.

- Must fail validation of a signed artifact with no key given.
  This test already, and still, passes.  It's just nice to have.

Changed TestValidate function to assert that the error is not nil
before asserting that it contains the expected text.
This makes a failure easier to understand when the error is
incorrectly nil, because it shows the line number of the assert
failure, instead of causing a panic before the assert is entered.
This happened to me when I ran the unit tests before fixing MEN-2155
and made it less obvious what was wrong.

Signed-off-by: Don Cross <cosinekitty@gmail.com>