Changes after review and full support for all fields in artifact v3.
Too many changes to count, but in general geared towards providing
full support for all headers and files in artifact format v3, not just
for rootfs but for the generic module image type, which will be used
for update modules.

Support has also been added for merging fields between original and
augmented updates, but this is missing the security feature (see
`PermittedAugmentedHeaders` in the update module spec) right now, and
everything can be merged, regardless of whether a field as been vetted
as safe or not, which can be used to override signatures. This should
be solved in the client by simply disallowing augmented fields until
the security support is in place.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>