Fix bug in artifact format regarding augmented indexing.
Because the augment types are in an indexed list, and they are not all
guaranteed to be present, we need to be explicit about missing entries
being represented by an empty string.

For now we simply don't write such an artifact, since it has some
tricky code paths in the writer. This part anyway needs to be cleaned
up later once we support writing more than one payload at a
time. However, it's important that the reader treats an augmented
payload whose type is empty as a missing augmented payload, otherwise
it opens up for security issues (an attacker can insert files into the
augmented part of a signed update, if that part was missing).

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>