Split ReaderArtifact() into Header and Data version.
This is required by the client because we need to write to the
database after reading the headers, but before reading the data.

For this we need to maintain more state between runs, so the outer tar
reader and the manifest were converted to class variables. The inner
tar readers for the headers and the data are still local and passed as
arguments.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>