Add Initialize() to UpdateStorer interface.
This is required by the client to set up the tree structure in update
modules. It cannot be done in PrepareStoreUpdate, because the client
needs to store the state to the database after the tree is created,
but before the update module is launched, and there is no room to do
that with PrepareStoreUpdate.

The reason it's important to store the state is so that we can
guarantee that the module gets to clean up if it has run at least one
state. If we didn't store this in the database early, we would not
know that we need to run it.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>