MEN-2486: Remove support for re-signing modified artifacts
It was not properly implemented; the "-k" parameter was being used as
the public key to verify the signature of the current Artifact and as
the private key as well to sign the modifyied Artifact.

In any case, it is better to completely remove this option and use the
mender-convert sign after modifications to produce the final Artifact

Removed the functionality and improved the test to actually check
something after the modifications.

Changelog: `mender-artifact modify` does not support anymore signing the
Artifact after the modification. Use `mender-convert sign` after the
modification to sign the Artifact.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>
(cherry picked from commit 19cbfed65f3cd900cc096523f02d9d93b21617dc)