MEN-2486: Clean up wrong/vain use of keys in copy commands
For all related copy commands (cp, cat, rm and install), the key
parameter was in one hand used used both as private and as public key;
but in the the other hand actually never parsable from the cli options.

This commits cleans it up, leaving the use of keys strictly to the
following commands: read and verify (using public key) plus write and
sign (using private key).

Internal functions repackArtifact and getCandidatesForModify in
artifacts.go are only used by modify and copy commans, so cleaning the
use of keys here also.

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>