Update vendored package crypto addind sys/windows dependency
The sys/windows package is being installed in build time, which will
potentially make it incompatible with the vendored version of crypto.
Furthermore, this pacakge seems now to be a hard dependency for crypto
so let's have both vendored to keep them always comparible.

This fixes the recent build break.

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>