MEN-2812: Enables Artifact Provides and Depends for write rootfs-img
Previously, the `write rootfs-image` command did not have the ability
to set Provides and Depends in the artifact. This was only enabled for
the `write module-image` command. Now the `rootfs-image` update can
also set Provides and Depends. However, please note that meta-data
and augmented Provides and Depends still are unsupported.

Changelog: commit

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>