MEN-2592: Refactor the artifact modification interface.
Changelog: Fix bug that destroys Artifact if any copy/modify command
is used on a non-rootfs-image Artifact.

In addition to fixing the bug above, this is done in preparation for
modifying arbitrary artifact attributes. There are several problems
with the current implementation which prompts this deep refactoring:

* The current implementation is centered around the `partition`
  struct, which is very partition focused, and does not map well to
  artifacts, especially non-rootfs-image artifacts

* Logic about operations on artifacts, raw images and partitioned
  images is spread out over several functions, making it very hard to
  make changes, and to keep the logic for one type in one place

* Since there is no central structure for representing each of the
  three types, artifacts, raw images and partitioned images, it is
  very hard to clean up properly, since you don't know where things
  have been stored, or even if anything has been stored

* The artifact modifier essentially throws out all information about
  the existing artifact and creates a brand new one from scratch. An
  artifact needs to retain all the metadata, so we need a structure
  that can hold all this

To solve this, the VPImage and VPFile interface were extended and
created, respectively. These mostly hide the details of the
modifications that happen to the actual image, and the modify code
need only worry about what to change, not how to change it.

The only exception is artifact attributes. For this the code has to
cast the interface to `ModImageArtifact` and make the necessary
adjustments. Casting seems like a reasonable tradeoff since it is the
only type of image to support such modifications. An interface would
not serve much purpose in this context.

We also added this next changelog entry, since it was extremely easy,
and a good convenvience feature:

Changelog: The `modify` subcommand has gained the `-k`/`--key`
argument to automatically sign the Artifact after modification.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>