MEN-2880: Make artifact install respect the given file permissions
The current Mender-Artifact install functionality had a bug in which the
temp-files it creates and uses for moving a file from the host to the target
with debugfs were created with the standard file permissions (0600). This did
pass test as only (0600) was tested. This fix creates a new temp-file, and
explicitly sets the file-permissions on it to match the one given on the command
line.

It has also added a couple of more test cases (0777, 0700) to make sure that the
same error as seen before does not slip unnoticed through the tests.

Changelog: Title

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>