MEN-2956: Write `rootfs_image_checksum` provides in rootfs-image
Previously the `write rootfs-image` command did not write the provided checksum
to the Artifact.

This commits makes sure the checksum is always provided when writing a
rootfs-image Artifact.

It also adds a command line flag under the `write` command, by the name of
`no-provides`, which disables the writing of this checksum, for compatibility
with older clients.

Changelog: Added the writing of the `rootfs_image_checksum` provide parameter as
a default to `rootfs-image` Artifacts. This means that now, the
`rootfs_image_checksum` will be written as a provide parameter to the Mender
client's database upon an update with the given Artifact. Please note that for
older clients (i.e. <= 2.1.x) this will not work, and the functionality should
be disabled by the user through the `--no-checksum-provide` flag when writing a
rootfs-image Artifact.

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>