Change the type of artifact.TypeInfoDepends and Provides
Previously artifact.TypeInfoDepends was of type pointer to map.

This commit removes the pointer, as every map in go is imiplicitly a pointer to
a map, and hence the extra pointer is unnecessary, and in fact has caused a lot
of fugly code in order to mitigate this poor design choice.

Changelog: None

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>