MEN-3276: Bugfix (cli/snapshot): signal-loop on ignored signals
If the forked ssh process received a signal which is ignored by default,
the signal is re-invoked using unix.Kill. This inevitably lead to an
invalid syscall instruction in case of SIGCHLD.

This commit makes SIGCHLD trigger terminal echo to be restored and
generates an error.

changelog: Bugfix: ignored signals no longer cause a signal-loop

Signed-off-by: Alf-Rune Siqveland <alf.rune@northern.tech>