[vendor] Added sha256-simd
This is a drop-in replacement for the sha-sum algorithm we are currently using,
taking advantage of the vector extensions present in modern CPU's if present.

This should, with little effort on our part, significantly speed up sha-sum
computations, which in the slowest part of handling the verification of the
artifacts at least.

Changelog: None

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>