Replaced 'crypto/sha256' with 'minio/sha256-simd' in checksum.go
This is to take advantage of the added package.

A quick profiling run on writing a file of 8GB resulted in:

 53.72s 15.51% 53.80%     53.72s 15.51%  crypto/sha256.block

And with simd extensions:

47.85s 14.31% 54.45%     47.85s 14.31%  github.com/mendersoftware/mender-artifact/vendor/github.com/minio/sha256-simd.blockAvx2

Giving approximately a 10 percent improvement in a function which takes up the
second most time during Artifact writing (the most time is taken by the
compression algorithm).

Changelog: None

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>