Updated the gzip tests to handle slightly different gzip reader
The pgzip package returns nil with a read of n-bytes > 0. In contrast to the
stdlib reader, which returned 'io.EOF' straight away. However, both are
perfectly valid implementations. According to the 'io.Reader' documentation:

When Read encounters an error or end-of-file condition after
successfully reading n > 0 bytes, it returns the number of
bytes read. It may return the (non-nil) error from the same call
or return the error (and n == 0) from a subsequent call.
An instance of this general case is that a Reader returning
a non-zero number of bytes at the end of the input stream may
return either err == EOF or err == nil. The next Read should
return 0, EOF.

Changelog: None

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>