Indexed the CLI commands by category
This should make it easier to distinguish the large number of CLI commands
depending on their intended usage.

The two categories added are:

* Artifact creation and validation
* Artifact modification

And should help to roughly set the commands apart depending on if they are
intended to work with a standard Artifact, either creating it or validating it.
The second category is intended for modification of already existing artifacts,
such as adding or removing files, signing or modifying the Artifact name.

Changelog: Commit

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>