Add a message reminding the user to check for missing Update-Module
A simple one-sentence addition to remind the user to check for a missing
update-module in case of an error.

This changes the error message from:

```
Artifact Payload type 'dir-overlay' is not supported by this Mender Client
```

To:

```
Artifact Payload type 'dir-overlay' is not supported by this Mender Client. Make
sure the Update Module is installed on the device.
```

Which makes for some less engineer specific information I think.

Changelog: None
Signed-off-by: Ole Petter <ole.orhagen@northern.tech>