Add test for tracking which flags are covered by certain commands.
This change was motivated by the fact that its really hard to track
which flags have been tested with which commands. For instance, after
a `write rootfs-image` flag is added, this flag needs to be tested
with both `dump` and `modify`, to make sure they take the changed
artifact into account.

We do this by scanning the command line arguments for some given
command, and then requiring that all flags are covered. There is no
way to automatically detect which flags are tested, so adding them to
the list of tested flags is still a manual process. But at least the
mechanism prevents you from adding a flag to the `write` subcommand
without being made aware that it also needs testing in other
subcommands.

This means also adding tests for some of the flags that were missing.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>