MEN-3479: Implement `clears_artifact_provides` field in Artifact format.
Changelog: Implement `clears_artifact_provides` field in Artifact
format. This field can be used to control how Artifacts modify the
record of existing software on the device. For example, a rootfs-image
update can erase the record of other software on the device, whereas a
single-file update can preserve the records. See the Mender
documentation for more information on how to use this, or refer to
`Documentation/artifact-format-v3.md` in the mender-artifact
repository for the reference.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>