Implement the scaffolding for reader and writer progress UX
This simply adds the framework needed in order to plug in a progress bar to
either the Artifact reader or the writer, as well as stage reporting while
writing the Artifact.

Notice that the reader and writer progressbar interfaces are not symmetric. This
is simply because there is no symmetry in our writer and reader code, and they
will have to be worked around in different ways. I am sorry.

Changelog: None
Signed-off-by: Ole Petter <ole.orhagen@northern.tech>