fix: Populate type in type-info for rootfs-image
Changelog: Fix regression which caused the "write rootfs-image"
command to produce an Artifact with empty "type" in type-info and
thus not adhering to the format specification. The same versions
that write such an Artifact will also accept it, but both issues
have been fixed: A corrupt Artifact is not written anymore, nor
will corrupt Artifacts be accepted. If you have produced an
Artifact using a version of mender-artifact prior to 4.3.1, you
may have to recreate it.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>
Signed-off-by: Vratislav Podzimek <vratislav.podzimek+auto-signed@northern.tech>