Fix writing of corrupt Artifact: type-info is missing type for rootfs-image.
Changelog: Fix regression which caused the "write rootfs-image"
command to produce a corrupt Artifact. The same versions that write
such an Artifact will also accept it, but both issues have been fixed:
A corrupt Artifact is not written anymore, nor will corrupt Artifacts
be accepted. If you have produced an Artifact using a version of
mender-artifact prior to 3.5.1, you may have to recreate it.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>