MEN-4502: Do not change the underlying Artifact unnecessarily
Previously the commands modifying an Artifact would always repack an Artifact,
no matter whether or not that modifications had actually been made to the
Artifact. As an example of this, if you had a signed Artifact compressed with
lzma, running `mender-artifact cat <artifact>:/<path-to-file>` would then cat
the file, and repack the Artifact with the standard compression, which is
`gzip`. Along the way the signature would also be lost.

This fix adds the following changes to the tooling:

* Modified images are no longer repacked, unless the command run has changed the
underlying image. This means that cat and copying out of an image will keep your
image intact. While copying into, installing, and removing files from the image
will repack the image.

* If an image is modified, and needs to be repacked, the existing compression
will be respected when repacking. The only exception is the `--compression` flag
for `mender-artifact modify` which can override the existing compression when repacking.

* `mender-artifact {cat,install,cp,rm}` do not respect the `--compression` flag,
but rather prints a warning, that the flag is ignored. If you want to change the
compression of your Artifact, run `mender-artifact modify <Artifact>
--compression <type>`

Changelog: Commit
Signed-off-by: Ole Petter <ole.orhagen@northern.tech>