Remove the '--compression' flag as global in the helptext
The flag was never really global anyways.

This commit removes it from the global option, and adds a note notifying the
user that the flag functions as a global flag, but only for certain commands.

Changelog: Add a note about the proper usage of the 'compression' flag in the
global help text.
Signed-off-by: Ole Petter <ole.orhagen@northern.tech>