Log: Notify that it is the script name which is wrong
It is easy to think that there is something inherently wrong with the script.

Make sure that the user understands it is the naming which is wrong, and add a
description and an example of the naming scheme expected.

Changelog: In case of a user trying to add a script with an invalid name, the
error message now says just so: Invalid script name, instead of simply: Invalid script.
Signed-off-by: Ole Petter <ole.orhagen@northern.tech>