fix: handle sparse partitioned SD-images
With recent changes to our image generation, the 'passive' partition is left
unpopulated by our image generation. This left the artifact modification
commands, like modify, cp, cat, etc in the dark, as they assumed both partitions
to be populated, and in fact did miror all changes to one rootfs to both.

This change adds a simple filter for the rootfs partition it modifies.

NOTE:

the heuristic used for detecting a sparse partition is simply looking for
the root filesystem directory.

Changelog: The mender-artifact modification commands (modify, cp, cat, rm) now
handles sparse partitions in the SD-images it modifies.

Ticket: MEN-5462

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>