chore: Switch mender-artifact to ubuntu:22 based builder.
This is done in order to link to libssl.so.3 instead of libssl.so.1.1.
Unfortunately ubuntu < 22 and ubuntu >= 22 are completely incompatible
in this regard, so we can't build one single binary which will work on
both of them.

We will solve this with a more generic distro package approach later.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>