fix: remove double hashing before signing and verifying using openssl
The openssl functions SignPKCS1v15 and VerifyPKCS1v15 accepts the
hashing function as first argument (openssl.SHA256_Method in our case)
and perform the hashing of the message internally. This differs from the
golang stdlib's crypto implementation, which requires the caller to pass
in a precomputed hash and the method argument is a flag used to
correctly detect the hash's size.

This commit removes the double hashing, which led to wrong signatures.

Changelog: None
Ticket: MEN-5759

Signed-off-by: Fabio Tranchitella <fabio.tranchitella@northern.tech>