ci: Publish Mac OS (and Windows) binaries on Git tags
Historically the publishing of binaries from tags was delegated to
`mender-qa` repository. This was removed some time ago in the context of
switching from binaries to packages but the Mac OS (and Windows)
binaries got lost in the process.

See:
* https://github.com/mendersoftware/mender-qa/commit/67b1b6bcd66af18b5b52814fde30ac3656473f5b

Repurpose the existing CI job here that was publishing the binaries for
branches to take care of the tags too, modernizing a bit the CI spec
along the way.

Ticket: MEN-8269

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>