fix: mender-artifact does not reenable echo on ssh error
Original idea was to handle reenabling echo thanks to EchoSigHandler, but it was failing sometimes.
The issue was that EchoSigHandler, when finishing its execution, sends an error to errChan. We need
to wait on errChan until we get something on it - meaning that the EchoSigHandler finished,
and we can safely exit the application.

However, before my changes, if there was an error returned before reaching
if s.sigChan != nil { signal.Stop(s.sigChan) s.cancel() if err := <-s.errChan; err != nil { return err } }

So, for example, here:
_, err = recvSnapshot(f, command.Stdout) if err != nil { _ = command.Cmd.Process.Kill() return "", err }

Then we never waitied on errChan. Application was closed, EchoSigHandler never got the chance to
reenable echo (sometimes it did, sometimes it didn't), and we were left with echo disabled.

Adding a separate function that waits for errChan, and deferring it, we have the problem solved.

Ticket: MEN-8428
Changelog: Make mender-artifact corretly reenable echo on any error

Signed-off-by: Michal Kopczan <michal.kopczan@northern.tech>